+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.3.0
+- Output of flowcharts:
+  "--flowchartout" and "--flowchartaddresses: Output flow-chart for a particular address (subroutine).
 - "--dot..." renamed to "--callgraph...".
-- "--flowchartout": Output flow-chart for a particular address (subroutine).
 
 
 ## 1.2.0